Lots of changes: o removed some old cvs log messages o some classes without member functions are in datatypes.py o movie_xml.parse now returns an object of MovieInformation instead of   a long list o mplayer_options now works better for options on cd/dvd/vcd o you can disable -wid usage o mplayer can play movies as strings or as FileInformation objects with   mplayer_options
git-svn-id: svn://svn.freevo.org/freevo/trunk/freevo@991 cc3e1ea1-1e01-0410-8d68-8b121e83a9d5

+++ mplayer.py
@@ -9,91 +9,17 @@
 #
 # -----------------------------------------------------------------------
 # $Log$
-# Revision 1.34  2002/09/18 18:42:19  dischi
-# Some small changes here and there, nothing important
+# Revision 1.35  2002/10/06 14:58:51  dischi
+# Lots of changes:
+# o removed some old cvs log messages
+# o some classes without member functions are in datatypes.py
+# o movie_xml.parse now returns an object of MovieInformation instead of
+#   a long list
+# o mplayer_options now works better for options on cd/dvd/vcd
+# o you can disable -wid usage
+# o mplayer can play movies as strings or as FileInformation objects with
+#   mplayer_options
 #
-# Revision 1.33  2002/09/14 16:49:20  dischi
-# Add support for hwac3. If MPLAYER_AO_HWAC3_DEV is set hwac3 will be
-# enabled for DVDs and VOB files
-#
-# Revision 1.32  2002/09/13 08:01:24  dischi
-# fix -cdrom-device for (S)VCD
-#
-# Revision 1.31  2002/08/21 04:58:26  krister
-# Massive changes! Obsoleted all osd_server stuff. Moved vtrelease and matrox stuff to a new dir fbcon. Updated source to use only the SDL OSD which was moved to osd.py. Changed the default TV viewing app to mplayer_tv.py. Changed configure/setup_build.py/config.py/freevo_config.py to generate and use a plain-text config file called freevo.conf. Updated docs. Changed mplayer to use -vo null when playing music. Fixed a bug in music playing when the top dir was empty.
-#
-# Revision 1.30  2002/08/18 06:10:58  krister
-# Converted tabs to spaces. Please use tabnanny in the future!
-#
-# Revision 1.29  2002/08/14 09:28:37  tfmalt
-#  o Updated all files using skin to create a skin object with the new
-#    get_singleton function. Please tell or add yourself if I forgot a
-#    place.
-#
-# Revision 1.28  2002/08/14 04:11:39  krister
-# Made the new X11 mplayer control feature resolution independent.
-#
-# Revision 1.27  2002/08/13 04:36:44  krister
-# Started adding testcode for controlling the MPlayer window under X11 (SDL).
-#
-# Revision 1.26  2002/08/13 02:01:02  krister
-# Made EXIT event stop the music player.
-#
-# Revision 1.25  2002/08/11 19:25:20  krister
-# Added more debug output.
-#
-# Revision 1.24  2002/08/08 03:15:57  krister
-# Tidied up some code.
-#
-# Revision 1.23  2002/08/05 00:50:33  tfmalt
-# o Split eventhandler into eventhandler_audio and eventhandler. This should
-#   make the code easier to maintain.
-# o Added start of eventhandling which manipulates the audio GUI (draw a
-#   volume meter when changing the volume etc.)
-#
-# Revision 1.22  2002/08/05 00:33:37  outlyer
-# Removed some debugging code I put in.
-#
-# Revision 1.21  2002/08/05 00:30:05  outlyer
-# Fixed the bug with mplayer detecting MP3 files as Encrypted VOBS. We now
-# force the demuxer for audio files, thereby fixing half of my MP3 collection.
-# Currently we only force demuxers for OGG and MP3 files, but check "demuxer.h"
-# in the mplayer source for a list of others. I don't believe they are
-# necessary though, as only MP3 has so many, many encoders with varying
-# output.
-#
-# While this alleviates the "urgency" of my plugin question, I still think
-# we should come up with a system of some kind.
-#
-# Revision 1.20  2002/08/04 16:49:20  dischi
-# DISPLAY toggles the mplayer OSD
-#
-# Revision 1.19  2002/08/03 18:55:44  outlyer
-# Last change to config file :)
-#
-# o You can now set the priority of the mplayer process via a nice setting
-# o This involves two lines in the config file: NICE and MPLAYER_NICE for the
-#       path to 'nice' and the actual numeric priority where '-10' is the
-#       default (high priority) set it to 0 for normal priority or +10 for
-#       low priority.
-#
-# Revision 1.18  2002/08/03 18:15:22  dischi
-# o added the patch from Thomas Malt for better audio control
-# o added support for the new freevo_config.py file
-#
-# Revision 1.17  2002/07/31 08:13:22  dischi
-# make screen black before mplayer starts in case mplayer doesn't cover
-# the hole screen
-#
-# Revision 1.15  2002/07/29 18:13:13  outlyer
-# Fixed "deja vu" of audio progressbar when playing movies after playing mp3s.
-#
-# Revision 1.14  2002/07/29 05:24:35  outlyer
-# Lots and lots of changes for new mplayer-based audio playing code.
-# o You'll need to modify your config file, as well as setup the new mplayer
-#   module by editing main.py
-# o This change includes Ogg Support, but that requires the ogg.vorbis
-#   module. If you don't want it, just don't install ogg.vorbis :)
 #
 #
 # -----------------------------------------------------------------------
@@ -134,6 +60,8 @@ import osd        # The OSD class, used to communicate with the OSD daemon
 import rc         # The RemoteControl class.
 import audioinfo  # This just for ID3 functions and stuff.
 import skin       # Cause audio handling needs skin functions.
+
+from datatypes import *
 
 DEBUG = 1
 TRUE  = 1
@@ -177,13 +105,21 @@ class MPlayer:
         self.thread.start()
         self.mode = None
                          
-    def play(self, mode, filename, playlist, repeat=0, mplayer_options=""):
+    def play(self, mode, file, playlist, repeat=0):
 
+        self.mode   = mode   # setting global var to mode.
+        self.repeat = repeat # Repeat playlist setting
+        mplayer_options = ''
+        
+        filename = file
+        if isinstance(filename, FileInformation):
+            mplayer_options = file.mplayer_options
+            mode = file.mode
+            filename = filename.file
+            
         if DEBUG:
             print 'MPlayer.play(): mode=%s, filename=%s' % (mode, filename)
             
-        self.mode   = mode   # setting global var to mode.
-        self.repeat = repeat # Repeat playlist setting
 
         if( (mode == 'video' or mode == 'audio') and
             not os.path.isfile(filename) ):
@@ -212,14 +148,9 @@ class MPlayer:
             # Mplayer command and standard arguments
             mpl += (' ' + config.MPLAYER_ARGS_MPG + ' -vo ' + config.MPLAYER_VO_DEV)
             
-            # Add special arguments for the hole playlist from the
-            # XML file
-            if mplayer_options:
-                mpl += (' ' + mplayer_options)
-                if DEBUG: print 'options, mpl = "%s"' % mpl
-
             # XXX Some testcode by Krister:
-            if os.path.isfile('./freevo_xwin') and osd.sdl_driver == 'x11':
+            if os.path.isfile('./freevo_xwin') and osd.sdl_driver == 'x11' and \
+               config.MPLAYER_USE_WID:
                 if DEBUG: print 'Got freevo_xwin and x11'
                 os.system('rm -f /tmp/freevo.wid')
                 os.system('./freevo_xwin  0 0 %s %s > /tmp/freevo.wid &' %
@@ -245,11 +176,6 @@ class MPlayer:
         elif mode == 'vcd':
             mpl += (' ' + config.MPLAYER_ARGS_VCD + ' -alang ' + config.DVD_LANG_PREF +
                     ' -vo ' + config.MPLAYER_VO_DEV)
-            # Add special arguments
-            if mplayer_options:
-                mpl += (' ' + mplayer_options)
-                if DEBUG: print 'options, mpl = "%s"' % mpl
-
             if config.DVD_SUBTITLE_PREF:
                 mpl += (' -slang ' + config.DVD_SUBTITLE_PREF)
             command = mpl % filename  # Filename is VCD chapter
@@ -259,19 +185,25 @@ class MPlayer:
             command = (mpl + " " + '-vo null ' + get_demuxer(filename) +
                        ' "' + filename + '"')
 
-        else:
+        elif mode == 'dvd':
             mpl += (' ' + config.MPLAYER_ARGS_DVD + ' -alang ' + config.DVD_LANG_PREF +
                     ' -vo ' + config.MPLAYER_VO_DEV)
             if config.DVD_SUBTITLE_PREF:
                 mpl += (' -slang ' + config.DVD_SUBTITLE_PREF)
+            command = mpl % str(filename)  # Filename is DVD chapter
+            
+        else:
+            print "Don't know what do play!"
             print "What is:      " + str(filename)
             print "What is mode: " + mode
             print "What is:      " + mpl
-            command = mpl % str(filename)  # Filename is DVD chapter
-            
+            return
 
+        # Add special arguments
+        if mplayer_options:
+            command += ' ' + mplayer_options
 
-        self.filename = filename 
+        self.file = file
         self.playlist = playlist
 
         # XXX A better place for the major part of this code would be
@@ -298,8 +230,6 @@ class MPlayer:
             osd.update()
 
             
-        self.mplayer_options = mplayer_options
-
         if DEBUG:
             print 'MPlayer.play(): Starting thread, cmd=%s' % command
             
@@ -367,18 +297,17 @@ class MPlayer:
                 menuwidget.refresh()
             else:
                 # Go to the previous movie in the list
-                pos = self.playlist.index(self.filename)
+                pos = self.playlist.index(self.file)
                 pos = (pos-1) % len(self.playlist)
-                filename = self.playlist[pos]
-                self.play(self.mode, filename, self.playlist, self.repeat,
-                          self.mplayer_options)
+                file = self.playlist[pos]
+                self.play(self.mode, file, self.playlist, self.repeat)
         elif event == rc.PLAY_END or event == rc.RIGHT:
             self.stop()
             if self.playlist == []:
                 rc.app = None
                 menuwidget.refresh()
             else:
-                pos = self.playlist.index(self.filename)
+                pos = self.playlist.index(self.file)
                 last_file = (pos == len(self.playlist)-1)
                 
                 # Don't continue if at the end of the list
@@ -388,9 +317,8 @@ class MPlayer:
                 else:
                     # Go to the next song in the list
                     pos = (pos+1) % len(self.playlist)
-                    filename = self.playlist[pos]
-                    self.play( self.mode, filename, self.playlist,
-                               self.repeat, self.mplayer_options )
+                    file = self.playlist[pos]
+                    self.play( self.mode, file, self.playlist, self.repeat)
         elif event == rc.VOLUP:
             print "Got VOLUP in mplayer!"
             # osd.popup_box('Volume shall come here')
@@ -438,11 +366,11 @@ class MPlayer:
                     menuwidget.refresh()
                 else:
                     # Go to the previous movie in the list
-                    pos = self.playlist.index(self.filename)
+                    pos = self.playlist.index(self.file)
                     pos = (pos-1) % len(self.playlist)
-                    filename = self.playlist[pos]
-                    self.play(self.mode, filename, self.playlist, self.repeat,\
-                              self.mplayer_options)
+                    file = self.playlist[pos]
+                    self.play(self.mode, file, self.playlist, self.repeat)
+
         elif event == rc.PLAY_END or event == rc.RIGHT:
             if event == rc.RIGHT and self.mode == 'dvdnav':
                 self.thread.app.write('L')
@@ -452,7 +380,7 @@ class MPlayer:
                     rc.app = None
                     menuwidget.refresh()
                 else:
-                    pos = self.playlist.index(self.filename)
+                    pos = self.playlist.index(self.file)
                     last_file = (pos == len(self.playlist)-1)
                 
                     # Don't continue if at the end of the list
@@ -463,8 +391,7 @@ class MPlayer:
                         # Go to the next song in the list
                         pos = (pos+1) % len(self.playlist)
                         filename = self.playlist[pos]
-                        self.play( self.mode, filename, self.playlist,
-                                   self.repeat, self.mplayer_options )
+                        self.play( self.mode, filename, self.playlist, self.repeat)
             
 # ======================================================================
 class MPlayerApp(childapp.ChildApp):
